+++ backend/app/Config/ConfigInterface.php
@@ -336,4 +336,9 @@ interface ConfigInterface
     public const FORCE_GITHUB_LINK = 'force_github_link';
     public const FORCE_MAIL_LINK = 'force_mail_link';
     public const FORCE_2FA = 'force_2fa';
+
+    /**
+     * Anti Adblocker.
+     */
+    public const ANTI_ADBLOCKER_ENABLED = 'anti_adblocker_enabled';
 }