+++ backend/app/Permissions.php
@@ -257,6 +257,10 @@ class Permissions
     /** Delete image reports */
     public const ADMIN_IMAGE_REPORTS_DELETE = 'admin.image.reports.delete';
 
+    // Adblocker Permissions
+    /** The permission that lets you bypass the antiadblocker! */
+    public const USER_PERMISSION_BYPASS_ADBLOCKER = 'user.permission.adblocker';
+
     /**
      * Returns all permission nodes with metadata.
      */
@@ -790,6 +794,12 @@ class Permissions
                 'value' => self::ADMIN_IMAGE_REPORTS_DELETE,
                 'category' => 'Admin Image Reports',
                 'description' => 'Delete image reports',
+            ],
+            [
+                'constant' => 'USER_PERMISSION_BYPASS_ADBLOCKER',
+                'value' => self::USER_PERMISSION_BYPASS_ADBLOCKER,
+                'category' => 'Adblocker',
+                'description' => 'The permission that lets you bypass the antiadblocker!',
             ],
         ];
     }