+++ backend/storage/migrations/2025-07-29.10.36-remove-old-tables.sql
@@ -1,0 +1,1 @@
+DROP TABLE `mythicaldash_analytics`, `mythicaldash_users_apikeys`;